Separating Colmap parsing functions into a new module
